requirements.txt
discord.py>=2.3.0
aiosqlite
fastapi
uvicorn
itsdangerous
python-dotenv
passlib[bcrypt]
jinja2
python-multipart
groq>=0.9.0
httpx>=0.27.0
PyNaCl>=1.5.0
yt-dlp>=2024.0.0
